Add Decimal library to make very small e values sortable

+++ MetaSanity/Parsers/blast_to_fasta.pyx
@@ -2,6 +2,7 @@
 from MetaSanity.Parsers.fasta_parser import FastaParser
 from libcpp.string cimport string
 import os
+from decimal import Decimal
 
 
 cdef extern from "Python.h":
@@ -9,7 +10,7 @@ cdef extern from "Python.h":
 
 
 def blast_to_fasta(str fasta_file, str blast_file, str out_file, int get_column = 0, tuple coords_columns = (2,3),
-                   float e_value = 1e-10, float pident = 98.5, tuple eval_pident_columns = (4,5), bint write_matches = True):
+                   double e_value = 1e-10, float pident = 98.5, tuple eval_pident_columns = (4,5), bint write_matches = True):
     """ Function will take the results of a blastx search and output corresponding
     fasta records from a file. Must provide the index of the column with id to get as int.
     Must also provide a tuple for the indices of the start/end coordinates, as well as for the indices of the evalue
@@ -40,7 +41,7 @@ def blast_to_fasta(str fasta_file, str blast_file, str out_file, int get_column 
         coords = (int(line[coords_columns[0]]), int(line[coords_columns[1]]))
         coords = tuple(sorted(coords))
         # Locate record based on column index and coords tuple passed
-        if e_value <= float(line[eval_pident_columns[0]]) and pident >= float(line[eval_pident_columns[1]]):
+        if e_value <= Decimal(line[eval_pident_columns[0]]) and pident >= float(line[eval_pident_columns[1]]):
             record_id = <string>PyUnicode_AsUTF8(line[get_column])
             record = fasta_records.get(record_id, None)
             if record:
